Enemies and weapons are separated in the objects folder
