+++ src/objects/Gunship.js
@@ -1,0 +1,32 @@
+import Phaser from 'phaser';
+import Entity from './Entities';
+import EnemyLaser from './enemyLaser';
+
+
+export default class GunShip extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, 'sprEnemy0', 'GunShip');
+    this.play('sprEnemy0');
+
+    this.body.velocity.y = Phaser.Math.Between(50, 100);
+
+    this.shootTimer = this.scene.time.addEvent({
+      delay: 1000,
+      callback() {
+        const laser = new EnemyLaser(this.scene, this.x, this.y);
+        laser.setScale(this.scaleX);
+        this.scene.enemyLasers.add(laser);
+      },
+      callbackScope: this,
+      loop: true,
+    });
+  }
+
+  onDestroy() {
+    if (this.shootTimer !== undefined) {
+      if (this.shootTimer) {
+        this.shootTimer.remove(false);
+      }
+    }
+  }
+}
