+++ src/objects/enemyLaser.js
@@ -1,0 +1,8 @@
+import Entity from './Entities';
+
+export default class EnemyLaser extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, 'sprLaserEnemy0');
+    this.body.velocity.y = 200;
+  }
+}
